% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binvar.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{Binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials, }{number of trials of the binomial distribution}

\item{prob, }{probability of succes of the binomial distriburion}
}
\value{
The binomial kurtosis of given number of trials and probability of success
}
\description{
Calculate the kurtosis for a binomial distribution
}
\examples{
bin_kurtosis(12, 0.7)
bin_kurtosis(5, 0.23)
}
