% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binvar.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials, }{number of trials of the binomial distribution}

\item{prob, }{probability of success of the binomial distribution}
}
\value{
An object of class "binvar"
}
\description{
This funciton creates a binomial variable object
}
\examples{
bin_variable(6, 0.5)
bin_variable(20, 0.5)
}
